requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Core scientific computing
numpy>=2.0.0

# LangChain ecosystem - FIXED VERSIONS for compatibility
langchain==0.3.7
langchain-core==0.3.21
langchain-community==0.3.7
langgraph==0.2.32

# OpenAI and AI dependencies
langchain-openai==0.2.8
openai>=1.35.0

# Experimental and specialized dependencies (use with caution)
langchain-experimental==0.3.1

# Cohere for reranking
langchain-cohere>=0.3.0
cohere>=4.0.0

# Vector stores and data processing
qdrant-client>=1.8.0
pandas>=2.0.0
pymupdf>=1.23.0

# Web search
tavily-python>=0.3.0

# Optional: Remove if not needed
# chromadb>=0.4.0